+++ wdsp/ssql.hpp
@@ -39,17 +39,17 @@ public:
     int size;                           // buffer size
     int rate;                           // sample-rate
     int rsize;                          // rate * time_to_fill_ring, e.g., 48K/s * 50ms = 2400
-    float fmax;                        // frequency (Hz) for full output, e.g., 2000 (Hz)
+    double fmax;                        // frequency (Hz) for full output, e.g., 2000 (Hz)
     float* in;                         // pointer to the intput buffer for ftov
     float* out;                        // pointer to the output buffer for ftov
     int* ring;                          // pointer to the base of the ring
     int rptr;                           // index into the ring
-    float inlast;                      // holds last sample from previous buffer
+    double inlast;                      // holds last sample from previous buffer
     int rcount;                         // count of zero-crossings currently in the ring
-    float div;                         // divisor for 'rcount' to produce output of 1.0 at 'fmax'
-    float eps;                         // minimum input change to count as a signal edge transition
+    double div;                         // divisor for 'rcount' to produce output of 1.0 at 'fmax'
+    double eps;                         // minimum input change to count as a signal edge transition
 
-    static FTOV* create_ftov (int run, int size, int rate, int rsize, float fmax, float* in, float* out);
+    static FTOV* create_ftov (int run, int size, int rate, int rsize, double fmax, float* in, float* out);
     static void destroy_ftov (FTOV *a);
     static void flush_ftov (FTOV *a);
     static void xftov (FTOV *a);
@@ -70,13 +70,13 @@ public:
     int rate;                           // sample rate
     int state;                          // state machine control
     int count;                          // count variable for raised cosine transitions
-    float tup;                         // time for turn-on transition
-    float tdown;                       // time for turn-off transition
+    double tup;                         // time for turn-on transition
+    double tdown;                       // time for turn-off transition
     int ntup;                           // number of samples for turn-on transition
     int ntdown;                         // number of samples for turn-off transition
     float* cup;                        // coefficients for up-slew
     float* cdown;                      // coefficients for down-slew
-    float muted_gain;                  // audio gain while muted; 0.0 for complete silence
+    double muted_gain;                  // audio gain while muted; 0.0 for complete silence
 
     float* b1;                         // buffer to hold output of dc-block function
     float* ibuff;                      // buffer containing only 'I' component
@@ -87,21 +87,21 @@ public:
     FTOV *cvtr;                         // pointer to F to V Converter data structure
     BQLP *filt;                         // pointer to Bi-Quad Low-Pass Filter data structure
     int ftov_rsize;                     // ring size for f_to_v converter
-    float ftov_fmax;                   // fmax for f_to_v converter
+    double ftov_fmax;                   // fmax for f_to_v converter
     // window detector
-    float wdtau;                       // window detector time constant
-    float wdmult;                      // window detector time constant multiplier
-    float wdaverage;                   // average signal value
-    float wthresh;                     // window threshold above/below average
+    double wdtau;                       // window detector time constant
+    double wdmult;                      // window detector time constant multiplier
+    double wdaverage;                   // average signal value
+    double wthresh;                     // window threshold above/below average
     // trigger
-    float tr_thresh;                   // trigger threshold:  100K/(100K+22K)=0.8197
-    float tr_tau_unmute;               // trigger unmute time-constant:  (100K||220K)*10uF = 0.6875
-    float tr_ss_unmute;                // trigger steady-state level for unmute:  100K/(100K+220K)=0.3125
-    float tr_tau_mute;                 // trigger mute time-constant:  220K*10uF = 2.2
-    float tr_ss_mute;                  // trigger steady-state level for mute:  1.0
-    float tr_voltage;                  // trigger voltage
-    float mute_mult;                   // multiplier for successive voltage calcs when muted
-    float unmute_mult;                 // multiplier for successive voltage calcs when unmuted
+    double tr_thresh;                   // trigger threshold:  100K/(100K+22K)=0.8197
+    double tr_tau_unmute;               // trigger unmute time-constant:  (100K||220K)*10uF = 0.6875
+    double tr_ss_unmute;                // trigger steady-state level for unmute:  100K/(100K+220K)=0.3125
+    double tr_tau_mute;                 // trigger mute time-constant:  220K*10uF = 2.2
+    double tr_ss_mute;                  // trigger steady-state level for mute:  1.0
+    double tr_voltage;                  // trigger voltage
+    double mute_mult;                   // multiplier for successive voltage calcs when muted
+    double unmute_mult;                 // multiplier for successive voltage calcs when unmuted
     int* tr_signal;                     // trigger signal, 0 or 1
 
     static SSQL* create_ssql (
@@ -110,15 +110,15 @@ public:
         float* in,
         float* out,
         int rate,
-        float tup,
-        float tdown,
-        float muted_gain,
-        float tau_mute,
-        float tau_unmute,
-        float wthresh,
-        float tr_thresh,
+        double tup,
+        double tdown,
+        double muted_gain,
+        double tau_mute,
+        double tau_unmute,
+        double wthresh,
+        double tr_thresh,
         int rsize,
-        float fmax
+        double fmax
     );
     static void destroy_ssql (SSQL *a);
     static void flush_ssql (SSQL *a);
@@ -128,9 +128,9 @@ public:
     static void setSize_ssql (SSQL *a, int size);
     // RXA Properties
     static void SetSSQLRun (RXA& rxa, int run);
-    static void SetSSQLThreshold (RXA& rxa, float threshold);
-    static void SetSSQLTauMute (RXA& rxa, float tau_mute);
-    static void SetSSQLTauUnMute (RXA& rxa, float tau_unmute);
+    static void SetSSQLThreshold (RXA& rxa, double threshold);
+    static void SetSSQLTauMute (RXA& rxa, double tau_mute);
+    static void SetSSQLTauUnMute (RXA& rxa, double tau_unmute);
 
 private:
     static void compute_ssql_slews(SSQL *a);
